devex: review feedback - normalize without the ./ in filenames to preserve suite-matching logic; fix instances of this.skip() not skipping; escaping the bar character (|) ended up being a bust so we'll delimit jest files with a comma instead

+++ cypress/deployed-and-local/integration/myAccount/verify-email-receipt.cy.ts
@@ -26,7 +26,14 @@ describe('Verify verification email', () => {
     cy.task('deleteAllItemsInEmailBucket', { bucketName });
   });
 
-  after(() => {
+  after(function () {
+    if (
+      getCypressEnv().isLocal ||
+      Cypress.env('MIGRATE') ||
+      Cypress.env('DISABLE_EMAILS')
+    ) {
+      return;
+    }
     cy.task('deleteAllItemsInEmailBucket', {
       bucketName,
       retries: 5,
